### 2024-01-02 #6180 @editorialbot
[PRE REVIEW]: NebulaSEM: A high-order discontinuous Galerkin spectral element code for atmospheric modeling
**Submitting author:** @dshawul (<a href="http://orcid.org/0000-0003-0671-2941">Daniel Abdi</a>)
**Repository:** https://github.com/dshawul/NebulaSEM
**Branch with paper.md** (empty if default branch): master
**Version:** 1.0
