### 2024-01-02 8e14990
Fix DOIs, include API doc links.
