+++ paper/paper.bib
@@ -18,7 +18,7 @@
     address = "Boston MA, USA",
     volume = "50",
     number = "13",
-    doi = "https://doi.org/10.1175/1520-0469(1993)050<1865:BCEWAS>2.0.CO;2",
+    doi = "10.1175/1520-0469(1993)050<1865:BCEWAS>2.0.CO;2",
     pages=  "1865 - 1873",
     url = "https://journals.ametsoc.org/view/journals/atsc/50/13/1520-0469_1993_050_1865_bcewas_2_0_co_2.xml"
 }
@@ -50,15 +50,18 @@
     year = {2011},
     doi = {10.1137/100791634}
 }
-
-@book{wallace2006,
-    author = {Wallace, J.M. and Hobbs, P.V.},
-    isbn = {9780127329512},
-    keywords = {Atmospheric Science},
-    publisher = {Elsevier Academic Press},
-    series = {International Geophysics Series},
-    title = {Atmospheric Science: An Introductory Survey},
-    year = 2006
+@article {stefanick1981,
+      author = "Michael  Stefanick",
+      title = "Space and Time Scales of Atmospheric Variability",
+      journal = "Journal of Atmospheric Sciences",
+      year = "1981",
+      publisher = "American Meteorological Society",
+      address = "Boston MA, USA",
+      volume = "38",
+      number = "5",
+      doi = "10.1175/1520-0469(1981)038<0988:SATSOA>2.0.CO;2",
+      pages = "988 - 1002",
+      url = "https://journals.ametsoc.org/view/journals/atsc/38/5/1520-0469_1981_038_0988_satsoa_2_0_co_2.xml"
 }
 
 @article{les, 
