### 2024-03-20 32752d3
improve explanation on how to run examples
