### 2024-04-02 962275a
Add notes about MPI+OpenMP
