### 2024-04-12 dshawul/NebulaSEM#4 @capitalaslash
algorithm in paper.md is cut
The comments in the algorithm section of the paper are cut.
Please use a smaller font and rephrase accordingly so that all the comments are in the margins.
