### 2024-04-12 691db6f
fix typos in paper.md (#5)
