### 2024-04-12 dshawul/solver#4 @dshawul
Hi @capitalaslash
I was not able to figure out how to decrease font size of the code block in "paper.md", but I have fixed
the issue by moving the comments inwards.

Thank you!
