### 2024-04-26 73f7d7a
Add basic test to catch missing grid file.
