### 2024-04-26 #6448 @esclapez
added two minor [PR7](https://github.com/dshawul/NebulaSEM/pull/7) and [PR8](https://github.com/dshawul/NebulaSEM/pull/8)
