### 2024-04-26 55a6aec
Fix test cavity path in README (#7)
* Fix path to test cavity-amr case.

* Add basic test to catch missing grid file.

* Exit with 0
