### 2024-04-26 7ab7ffe
Don't compute rho from p.
