+++ apps/euler/euler.cpp
@@ -58,7 +58,7 @@ void euler(std::istream& input) {
         ScalarCellField p("p",READWRITE);
         VectorCellField U("U",READWRITE);
         ScalarCellField T("T",READWRITE);
-        ScalarCellField rho("rho",READWRITE);
+        ScalarCellField rho("rho",WRITE);
         VectorCellField g(false);
 
         /*Read fields*/
@@ -92,7 +92,7 @@ void euler(std::istream& input) {
                     U[i][1] += (beta / (2 * PI)) * exp((1 - r[i]*r[i])/2.0) *  Mesh::cC[i][0];
                 }
                 p = pow(T + T0, p_gamma / (p_gamma - 1)) - P0;
-                rho = (P0 / (R*(T + T0))) * pow((p + P0) / P0, 1 / p_gamma);
+                rho = (P0 / (R*(T + T0))) * pow((p + P0) / P0, 1 / p_gamma) - (P0 / (R*T0));
             };
 
             /*choose init type*/
@@ -135,40 +135,23 @@ void euler(std::istream& input) {
             rho_ref = (P0 / (R*T0));
         }
 
-        if(ait.start()) {
-            p += p_ref;
-            Mesh::scaleBCs<Scalar>(p,p_ref,1.0);
-            applyExplicitBCs(p_ref,true);
-            applyExplicitBCs(p,true);
+        /* compute rho from p*/
+        p += p_ref;
+        Mesh::scaleBCs<Scalar>(p,p_ref,1.0);
+        applyExplicitBCs(p_ref,true);
+        applyExplicitBCs(p,true);
 
-            /*calculate rho*/
-            Mesh::scaleBCs<Scalar>(p,rho_ref,psi);
-            applyExplicitBCs(rho_ref,true);
+        /*calculate rho*/
+        Mesh::scaleBCs<Scalar>(p,rho_ref,psi);
+        applyExplicitBCs(rho_ref,true);
 
-            rho = (P0 / (R*(T + T0))) * pow(p / P0, 1 / p_gamma);
-            Mesh::scaleBCs<Scalar>(p,rho,psi);
-            applyExplicitBCs(rho,true);
-            rho -= rho_ref;
-            rho.write(0);
+        rho = (P0 / (R*(T + T0))) * pow(p / P0, 1 / p_gamma);
+        Mesh::scaleBCs<Scalar>(p,rho,psi);
+        applyExplicitBCs(rho,true);
+        rho -= rho_ref;
+        rho.write(0);
 
-            p -= p_ref;
-        } else {
-            rho += rho_ref;
-            Mesh::scaleBCs<Scalar>(rho,rho_ref,1.0);
-            applyExplicitBCs(rho_ref,true);
-            applyExplicitBCs(rho,true);
-
-            /*calculate pressure*/
-            Mesh::scaleBCs<Scalar>(rho,p_ref,1.0/psi);
-            applyExplicitBCs(p_ref,true);
-
-            p = P0 * pow((rho*(T+T0)*R) / P0, p_gamma);
-            Mesh::scaleBCs<Scalar>(rho,p,1.0/psi);
-            applyExplicitBCs(p,true);
-            p -= p_ref;
-
-            rho -= rho_ref;
-        }
+        p -= p_ref;
 
         /*compute total mass and energy*/
         if(ait.get_step() == 0)
