+++ src/field/field.h
@@ -434,17 +434,17 @@ class BaseField {
         //------------
         virtual void writeInternal(std::ostream&,IntVector*) = 0;
         virtual Int readInternal(std::istream&,Int = 0) = 0;
-        virtual void writeBoundary(std::ostream&) = 0;
+        virtual void writeBoundary(std::ostream&, Mesh::MeshObject* = 0) = 0;
         virtual void readBoundary(std::istream&) = 0;
         virtual void read(std::istream&) = 0;
-        virtual void write(std::ostream&, IntVector* = 0) = 0;
+        virtual void write(std::ostream&, IntVector* = 0, Mesh::MeshObject* = 0) = 0;
 
         virtual void writeInternal(Util::ofstream_bin&,IntVector*) = 0;
         virtual Int readInternal(Util::ifstream_bin&,Int = 0) = 0;
-        virtual void writeBoundary(Util::ofstream_bin&) = 0;
+        virtual void writeBoundary(Util::ofstream_bin&, Mesh::MeshObject* = 0) = 0;
         virtual void readBoundary(Util::ifstream_bin&) = 0;
         virtual void read(Util::ifstream_bin&) = 0;
-        virtual void write(Util::ofstream_bin&, IntVector* = 0) = 0;
+        virtual void write(Util::ofstream_bin&, IntVector* = 0, Mesh::MeshObject* = 0) = 0;
         //------------
         virtual void read(Int) = 0;
         virtual void write(Int, IntVector* = 0) = 0;
@@ -761,11 +761,11 @@ class MeshField : public BaseField
         template<typename Ts>
         void writeInternal_(Ts&,IntVector*);
         template<typename Ts>
-        void writeBoundary_(Ts&);
+        void writeBoundary_(Ts&, Mesh::MeshObject*);
         template<typename Ts>
         void read_(Ts&);
         template<typename Ts>
-        void write_(Ts&, IntVector*);
+        void write_(Ts&, IntVector*, Mesh::MeshObject*);
         //------------
     public:
         ACCESS       access;
@@ -1007,16 +1007,16 @@ class MeshField : public BaseField
         Int readInternal(std::istream& is,Int offset = 0) override { return readInternal_(is,offset); }
         void readBoundary(std::istream& is) override { readBoundary_(is); }
         void writeInternal(std::ostream& os,IntVector* cMap) override { writeInternal_(os, cMap); };
-        void writeBoundary(std::ostream& os) override {writeBoundary_(os); }
+        void writeBoundary(std::ostream& os, Mesh::MeshObject* pmesh = 0) override {writeBoundary_(os, pmesh); }
         void read(std::istream& os) override {read_(os); }
-        void write(std::ostream& os, IntVector* cMap) override {write_(os, cMap); }
+        void write(std::ostream& os, IntVector* cMap, Mesh::MeshObject* pmesh = 0) override {write_(os, cMap, pmesh); }
 
         Int readInternal(Util::ifstream_bin& is,Int offset = 0) override { return readInternal_(is,offset); }
         void readBoundary(Util::ifstream_bin& is) override { readBoundary_(is); }
         void writeInternal(Util::ofstream_bin& os,IntVector* cMap) override { writeInternal_(os, cMap); };
-        void writeBoundary(Util::ofstream_bin& os) override {writeBoundary_(os); }
+        void writeBoundary(Util::ofstream_bin& os, Mesh::MeshObject* pmesh = 0) override {writeBoundary_(os, pmesh); }
         void read(Util::ifstream_bin& os) override {read_(os); }
[…]
