### 2024-04-27 a14320d
Bug fix in merging text files.
