### 2024-04-27 e9780b7
Rewrite AmrTree on decomposition.
