### 2024-04-28 f155810
Modifying some test settings.
