### 2024-04-28 d8bf9d1
Add github actions test for parallel runs.
