+++ README.md
@@ -1,6 +1,7 @@
 [![NebulaSEM](https://github.com/dshawul/NebulaSEM/actions/workflows/NebulaSEM.yml/badge.svg)](https://github.com/dshawul/NebulaSEM/actions/workflows/NebulaSEM.yml)
 [![APIdoc](https://github.com/dshawul/NebulaSEM/actions/workflows/APIdoc.yml/badge.svg)](https://github.com/dshawul/NebulaSEM/actions/workflows/APIdoc.yml)
 [![JOSS](https://github.com/dshawul/NebulaSEM/actions/workflows/paper.yml/badge.svg)](https://github.com/dshawul/NebulaSEM/actions/workflows/paper.yml)
+[![DOI](https://zenodo.org/badge/28114781.svg)](https://zenodo.org/doi/10.5281/zenodo.11088209)
 
 ## NebulaSEM
 
