### 2024-04-29 #6448 @dshawul
@observingClouds Thank you! I have tried to address the author's checklists as follows
- The author and affiliations are correct
- I have created a release "NebulaSEM v2.0"
- Archived the release on Zenodo with the DOI
[![DOI](https://zenodo.org/badge/28114781.svg)](https://zenodo.org/doi/10.5281/zenodo.11088209)
- The Zenodo archive author list is different in that it contains two github accounts of mine, and the two JOSS reviewers who contributed to the repo. I suppose that is Ok so just letting you know.
- The license on Zenodo says "Creative Commons License" while the github one has "BSD-2-clause simplified license".
Is that Ok?

Sorry I am new to this.
