### 2024-04-29 #6448 @danielabdi-noaa
@observingClouds Sorry, I was not aware that you are able to edit the archive. I have now fixed both the "LICENSE" issue and the "Author's list" issue. Let me know if there are things that I missed.
