### 2024-04-30 #6448 @observingClouds
Thanks @dshawul, no worries. Could you also adjust the title of the archived release? We prefer to have the JOSS release having the same title as the paper, i.e. "NebulaSEM: A high-order discontinuous Galerkin
spectral element code for atmospheric modeling". Thank you.
