+++ paper/paper.bib
@@ -11,7 +11,7 @@
 
 @article {robert1993,
     author = "Andre  Robert",
-    title = "Bubble Convection Experiments with a Semi-implicit Formulation of the Euler Equations",
+    title = {{Bubble Convection Experiments with a Semi-implicit Formulation of the Euler Equations}},
     journal = "Journal of Atmospheric Sciences",
     year = "1993",
     publisher = "American Meteorological Society",
@@ -41,8 +41,8 @@
 }
 @ARTICLE{p4est,
     author = {Carsten Burstedde and Lucas C. Wilcox and Omar Ghattas},
-    title = {{\texttt{p4est}}: Scalable Algorithms for Parallel Adaptive Mesh
-           Refinement on Forests of Octrees},
+    title = {{{\texttt{p4est}}: Scalable Algorithms for Parallel Adaptive Mesh
+           Refinement on Forests of Octrees}},
     journal = {SIAM Journal on Scientific Computing},
     volume = {33},
     number = {3},
@@ -52,7 +52,7 @@
 }
 @article {stefanick1981,
       author = "Michael  Stefanick",
-      title = "Space and Time Scales of Atmospheric Variability",
+      title = {{Space and Time Scales of Atmospheric Variability}},
       journal = "Journal of Atmospheric Sciences",
       year = "1981",
       publisher = "American Meteorological Society",
@@ -86,7 +86,7 @@
 }
 @article{abdi8,
     author = {Daniel S Abdi and Lucas C Wilcox and Timothy C Warburton and Francis X Giraldo},
-    title = {A {GPU}-accelerated continuous and discontinuous Galerkin non-hydrostatic atmospheric model},
+    title = {A {GPU}-accelerated continuous and discontinuous {Galerkin} non-hydrostatic atmospheric model},
     journal = {The International Journal of High Performance Computing Applications},
     volume = {22},
     issue = {1},
