### 2024-04-30 #6448 @kthyng
Congratulations on your new publication @dshawul! Many thanks to @observingClouds and to reviewers @capitalaslash and @esclapez for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
